[REFACTOR] 문서 구조 변경
