+++ js/header/header.js
@@ -1,6 +1,6 @@
 import { refreshPage } from "./refreshPage.js";
 import { getDate } from "./getDate.js";
-import { REFRESH_DATE_TIME } from "../constant.js";
+import { REFRESH_DATE_TIME } from "../utils/constant.js";
 
 function initHeader() {
   getDate(".date", REFRESH_DATE_TIME);
