--- js/main/listView/makeListView.js
@@ -1,7 +1,0 @@
-import { makeCategory, transformMainNews } from "./handleNewsData.js";
-
-function makeListView() {
-  makeCategory();
-  transformMainNews();
-}
-export { makeListView };
