+++ js/main/listView/renderListView.js
@@ -1,20 +1,22 @@
 import { fetchData } from "../../utils/fetchData.js";
 import { viewOption } from "../../store/store.js";
 import { getState } from "../../store/observer.js";
-import { manipulateNewsByCategory } from "./manipulateNewsByCategory.js";
-import { manipulateNewsByPress } from "./manipulateNewsByPress.js";
-import { makeListView } from "./makeListView.js";
-import { ALL_PRESS } from "../../constant.js";
+import {
+  manipulateNewsByCategory,
+  manipulateNewsByPress,
+} from "./manipulateNews.js";
+import { ALL_PRESS } from "../../utils/constant.js";
+import { makeCategory, transformMainNews } from "./handleNewsData.js";
 
 async function renderListView() {
   const news = await fetchData(".././assets/news.json");
   if (getState(viewOption) === ALL_PRESS) {
     manipulateNewsByCategory(news);
-    makeListView();
   } else {
     manipulateNewsByPress(news);
-    makeListView();
   }
+  makeCategory();
+  transformMainNews();
 }
 
 export { renderListView };
